Refactor: Remove common route logic from BottomNavigationBar and bottom bar from Details and Search screens
This commit removes the `itemsInCommon` logic from `BottomNavigationBar.kt` which was used to highlight the "More" icon when the user was on the `Search` screen.

Additionally, the `BottomNavigationBar` has been removed from the `Details.kt` and `Search.kt` screens as it was not intended to be displayed on these screens.