Refactor: Improve navigation logic in More screen
This commit refactors the `More.kt` screen to improve the navigation logic when an option is clicked.

The `clickable` modifier now checks the selected option's text. Navigation to the `Search` screen is triggered only if the clicked item corresponds to the "Search" option. The "Favourites" option currently has no action associated with it.